move siri_model into a module and split it by api
the model was becoming too big, split it by siri's reponse
