+++ src/siri_lite/service_delivery.rs
@@ -1,44 +1,4 @@
-use navitia_model::collection::Idx;
-use navitia_model::objects::StopPoint;
-use navitia_model::Model;
-
-#[derive(Debug)]
-pub struct DateTime(pub chrono::NaiveDateTime);
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ErrorCondition {}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct Line {
-    pub line_ref: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Location {
-    pub longitude: f64,
-    pub latitude: f64,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct AnnotatedStopPoint {
-    pub stop_point_ref: String,
-    pub stop_name: String,
-    pub lines: Vec<Line>,
-    pub location: Location,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct StopPointsDelivery {
-    pub version: String,
-    pub response_time_stamp: String,
-    pub status: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error_condition: Option<ErrorCondition>,
-    pub annotated_stop_point: Vec<AnnotatedStopPoint>,
-}
+use crate::siri_lite::DateTime;
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -152,72 +112,4 @@ pub struct ServiceDelivery {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub request_message_ref: Option<String>, // Note: this is mandatory for idf profil
     pub stop_monitoring_delivery: Vec<StopMonitoringDelivery>,
-}
-
-#[derive(Debug, Default, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct Siri {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub stop_points_delivery: Option<StopPointsDelivery>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub service_delivery: Option<ServiceDelivery>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct SiriResponse {
-    pub siri: Siri,
-}
-
-impl AnnotatedStopPoint {
-    pub fn from(stop_point_idx: Idx<StopPoint>, model: &Model) -> Self {
-        let lines = model
-            .get_corresponding_from_idx(stop_point_idx)
-            .into_iter()
-            .map(|route_id| Line {
-                line_ref: model.routes[route_id].id.clone(),
-            })
-            .collect();
-
-        let sp = &model.stop_points[stop_point_idx];
-
-        Self {
-            stop_point_ref: sp.id.clone(),
-            stop_name: sp.name.clone(),
-            lines,
-            location: Location {
-                longitude: sp.coord.lon,
-                latitude: sp.coord.lat,
-            },
-        }
-    }
-}
-
-impl std::string::ToString for DateTime {
-    fn to_string(&self) -> String {
-        self.0.format("%Y-%m-%dT%H:%M:%S").to_string()
-    }
-}
-
-impl serde::Serialize for DateTime {
-    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
-    where
-        S: ::serde::Serializer,
-    {
-        serializer.serialize_str(&self.to_string())
-    }
-}
-
-impl<'de> ::serde::Deserialize<'de> for DateTime {
-    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
-    where
-        D: ::serde::Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        Ok(DateTime(
-            chrono::NaiveDateTime::parse_from_str(&s, "%Y-%m-%dT%H:%M:%S").map_err(|e| {
-                serde::de::Error::custom(format!("datetime format not valid: {}", e))
-            })?,
-        ))
-    }
 }
