+++ src/siri_lite/shared.rs
@@ -1,0 +1,32 @@
+
+#[derive(Debug)]
+pub struct DateTime(pub chrono::NaiveDateTime);
+
+impl std::string::ToString for DateTime {
+    fn to_string(&self) -> String {
+        self.0.format("%Y-%m-%dT%H:%M:%S").to_string()
+    }
+}
+
+impl serde::Serialize for DateTime {
+    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
+    where
+        S: ::serde::Serializer,
+    {
+        serializer.serialize_str(&self.to_string())
+    }
+}
+
+impl<'de> ::serde::Deserialize<'de> for DateTime {
+    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
+    where
+        D: ::serde::Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+        Ok(DateTime(
+            chrono::NaiveDateTime::parse_from_str(&s, "%Y-%m-%dT%H:%M:%S").map_err(|e| {
+                serde::de::Error::custom(format!("datetime format not valid: {}", e))
+            })?,
+        ))
+    }
+}
