+++ tests/stop_monitoring_test.rs
@@ -1,6 +1,6 @@
 use actix_web::http;
 use actix_web::HttpMessage;
-use transpo_rt::siri_model::{DateTime, SiriResponse};
+use transpo_rt::siri_lite::{DateTime, SiriResponse};
 use transpo_rt::transit_realtime;
 mod utils;
